fraud_genai_project/requirements.txt
transformers==4.39.3
sentence-transformers==2.6.1
faiss-cpu==1.7.4
faker==37.1.0